+++ store/src/Haskoin/Store.hs
@@ -3,32 +3,27 @@ module Haskoin.Store
     , StoreConfig(..)
     , StoreEvent(..)
     , withStore
-
     , module Haskoin.Store.BlockStore
     , module Haskoin.Store.Web
     , module Haskoin.Store.Database.Reader
     , module Haskoin.Store.Data
-
       -- * Cache
-    , CacheConfig (..)
+    , CacheConfig(..)
     , CacheT
-    , CacheError (..)
+    , CacheError(..)
     , withCache
     , connectRedis
     , isInCache
     , evictFromCache
-
       -- * Store Reader
-    , StoreRead (..)
-    , Limit
-    , Offset
-
+    , StoreRead(..)
+    , Limits(..)
+    , Start(..)
       -- * Useful Fuctions
     , getTransaction
     , blockAtOrBefore
-
       -- * Other Data
-    , PubExcept (..)
+    , PubExcept(..)
     ) where
 
 import           Haskoin.Store.BlockStore